feat: no se what I did, these changes have been on my pc for like a year now
