+++ kigoutchi.code-workspace
@@ -1,0 +1,26 @@
+{
+	"folders": [
+		{
+			"path": "."
+		}
+	],
+	"settings": {
+		"rust-analyzer.linkedProjects": [
+			".\\Cargo.toml"
+		],
+		"workbench.colorTheme": "One Dark Pro",
+		"workbench.preferredDarkColorTheme": "One Dark Pro",
+		"workbench.iconTheme": "bearded-icons",
+		"editor.fontSize": 18,
+		"editor.lineHeight": 30,
+		"editor.fontFamily": "JetBrains Mono, monospace",
+		"cSpell.language": "en,es",
+		"cSpell.words": [
+			"kato",
+			"meaw",
+			"recv",
+			"rowstr",
+			"vect"
+		],
+	}
+}
