+++ src/pet/mod.rs
@@ -1,0 +1,6 @@
+pub mod pet;
+
+pub use pet::Pet;
+pub use pet::PetStatus;
+pub use pet::CatGrid;
+pub use pet::CatSize;
