+++ src/utils/mod.rs
@@ -1,0 +1,7 @@
+pub mod grid;
+pub mod utils;
+// pub mod charinfo;
+
+pub use grid::Grid;
+pub use utils::CharInfo;
+pub use utils::DoubleCharInfo;
