--- tamagotchi.code-workspace
@@ -1,11 +1,0 @@
-{
-	"folders":[
-		{
-			"path":"."
-		}
-	],
-	"settings":{
-        "workbench.iconTheme": "easy-icons",
-		"workbench.colorTheme": "SynthWave '84",
-    }
-}
